Refactored the class name of all package. issue no #11
