+++ src/study/graphTheory/GraphTheoryUtility.kt
@@ -1,0 +1,9 @@
+package study.graphTheory
+
+object GraphTheoryUtility {
+
+    /**
+     * 引数で指定された数値の総ペア数を求める。
+     */
+    fun calcPairCombinationNum(num: Long) = (num * (num - 1)) / 2
+}
